Added support for all of the capture option classes to GrabzItBundle
I just came across your awesome component!

I thought it would be great if I could add support for all of the different option classes and then if/when released it will help all Symfony users. We plan to link to it from our website as well giving you credit of course for creating such a cool Symfony bundle!

If you have anything specific you want me to link to please mention it!